Make CKRelativeDimension ctors inlineable
Helps with code size.
