+++ ComponentKit/Core/CKDimension.h
@@ -12,6 +12,15 @@
 
 #import <UIKit/UIKit.h>
 
+#import <ComponentKit/CKAssert.h>
+#import <ComponentKit/CKSizeRange.h>
+#import <ComponentKit/ComponentLayoutContext.h>  // Used by CKCAssertPositiveReal.
+
+#define CKCAssertPositiveReal(description, num) \
+  CKCAssert(num >= 0 && num < CGFLOAT_MAX, @"%@ must be a real positive integer.\n%@", description, CK::Component::LayoutContext::currentStackDescription())
+#define CKCAssertInfOrPositiveReal(description, num) \
+  CKCAssert(isinf(num) || (num >= 0 && num < CGFLOAT_MAX), @"%@ must be infinite or a real positive integer.\n%@", description, CK::Component::LayoutContext::currentStackDescription())
+
 /**
  A dimension relative to constraints to be provided in the future.
  A RelativeDimension can be one of three types:
@@ -53,33 +62,16 @@ private:
     POINTS,
     PERCENT,
   };
-  CKRelativeDimension(Type type, CGFloat value);
+  CKRelativeDimension(Type type, CGFloat value)
+    : _type(type), _value(value)
+  {
+    if (type == Type::POINTS) {
+      CKCAssertPositiveReal(@"Points", value);
+    }
+  }
+
   Type _type;
   CGFloat _value;
-};
-
-/** Expresses an inclusive range of sizes. Used to provide a simple constraint to component layout. */
-struct CKSizeRange {
-  CGSize min;
-  CGSize max;
-
-  /** The default constructor creates an unconstrained range. */
-  CKSizeRange() : CKSizeRange({0,0}, {INFINITY, INFINITY}) {}
-
-  CKSizeRange(const CGSize &min, const CGSize &max);
-
-  /** Clamps the provided CGSize between the [min, max] bounds of this SizeRange. */
-  CGSize clamp(const CGSize &size) const;
-
-  /**
-   Intersects another size range. If the other size range does not overlap in either dimension, this size range
-   "wins" by returning a single point within its own range that is closest to the non-overlapping range.
-   */
-  CKSizeRange intersect(const CKSizeRange &other) const;
-
-  bool operator==(const CKSizeRange &other) const;
-  NSString *description() const;
-  size_t hash() const;
 };
 
 /** Expresses a size with relative dimensions. */
